fix: route library search through bucket sessions
